+++ dash-spv-masternode-processor/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "dash-spv-masternode-processor"
-version = "0.4.8"
+version = "0.4.9"
 description = "Library for processing masternodes and quorums (SPV)"
 
 edition = "2021"
@@ -24,23 +24,24 @@ std = ["secp256k1/std", "bitcoin_hashes/std"]
 #no-std = ["hashbrown", "core2/alloc", "bitcoin_hashes/alloc"]
 
 [dependencies]
-byte = "0.2"
-libc = "0.2.132"
+bip38 = { git = "https://github.com/pankcuf/bip38", rev = "87abd21" }
 bitcoin_hashes = { version = "0.11.0", default-features = false }
 bitcoinconsensus = { version = "0.19.0-3", optional = true }
 blake3 = "1.3.2"
 bls-signatures = { git = "https://github.com/dashevo/bls-signatures", rev = "a5589e6", features = ["legacy", "bip32", "apple"]  }
-ed25519-dalek = "2.0.0-pre.0"
-secp256k1 = { version = "0.26.0", features = [ "recovery", "rand-std", "bitcoin_hashes" ] }
+byte = "0.2"
 core2 = { version = "0.4.0", optional = true, default-features = false }
-rs-x11-hash = "0.1.8"
-zeroize = "1.5.7"
-bip38 = { git = "https://github.com/pankcuf/bip38", rev = "87abd21" }
-log = { version = "0.4.*", features = ["std"] }
-simplelog = "0.12.1"
 dirs-next = "2.0.0"
+ed25519-dalek = "2.0.0-pre.0"
+libc = "0.2.132"
+log = { version = "0.4.*", features = ["std"] }
+rs-x11-hash = "0.1.8"
+secp256k1 = { version = "0.26.0", features = [ "recovery", "rand-std", "bitcoin_hashes" ] }
 serde = { version = "1.0.144", features = ["derive"] }
 serde_json = "1.0.85"
+simplelog = "0.12.1"
+zeroize = "1.5.7"
+
 [dev-dependencies]
 base64 = "0.21.0"
 reqwest = { version = "0.11", features = ["blocking", "json"] }
