+++ dash-spv-masternode-processor/src/bindings/keys.rs
@@ -6,9 +6,8 @@ use byte::BytesExt;
 use secp256k1::Scalar;
 use crate::chain::bip::bip32;
 use crate::chain::bip::bip38::BIP38;
-use crate::chain::common::chain_type::IHaveChainSettings;
+use crate::chain::common::{ChainType, IHaveChainSettings};
 use crate::chain::derivation::{BIP32_HARD, IndexPath};
-use crate::common::ChainType;
 use crate::consensus::Encodable;
 use crate::crypto::byte_util::{AsBytes, clone_into_array, ConstDecodable, Reversable, Zeroable};
 use crate::crypto::{UInt160, UInt256, UInt384, UInt512, UInt768};
