+++ dash-spv-masternode-processor/src/bindings/masternode.rs
@@ -2,8 +2,7 @@ use std::ptr::null_mut;
 use std::slice;
 use byte::BytesExt;
 use crate::{models, types};
-use crate::chain::common::chain_type::IHaveChainSettings;
-use crate::common::{ChainType, LLMQType};
+use crate::chain::common::{ChainType, IHaveChainSettings, LLMQType};
 use crate::consensus::encode;
 use crate::crypto::{UInt256, byte_util::{BytesDecodable, ConstDecodable}};
 use crate::ffi::{boxer::{boxed, boxed_vec}, ByteArray, from::FromFFI, to::ToFFI};
