+++ dash-spv-masternode-processor/src/chain/common/mod.rs
@@ -1,3 +1,5 @@
 pub mod chain_type;
+pub mod llmq_type;
 
-pub use self::chain_type::ChainType;
+pub use self::chain_type::{ChainType, IHaveChainSettings, DevnetType};
+pub use self::llmq_type::{DKGParams, LLMQParams, LLMQType};
