+++ dash-spv-masternode-processor/src/common/mod.rs
@@ -1,15 +1,12 @@
 pub mod block;
 pub mod llmq_snapshot_skip_mode;
-pub mod llmq_type;
 pub mod llmq_version;
 pub mod merkle_tree;
 pub mod socket_address;
 pub mod masternode_type;
 
 pub use self::block::Block;
-pub use crate::chain::common::chain_type::ChainType;
 pub use self::llmq_snapshot_skip_mode::LLMQSnapshotSkipMode;
-pub use self::llmq_type::{DKGParams, LLMQParams, LLMQType};
 pub use self::llmq_version::LLMQVersion;
 pub use self::masternode_type::MasternodeType;
 pub use self::merkle_tree::MerkleTree;
