+++ dash-spv-masternode-processor/src/ffi/from.rs
@@ -1,6 +1,6 @@
 use std::collections::BTreeMap;
 use std::slice;
-use crate::{common, models, tx, types};
+use crate::{chain, common, models, tx, types};
 use crate::consensus::encode;
 use crate::crypto::{byte_util::Reversable, UInt128, UInt160, UInt256, UInt384, UInt768};
 use crate::ffi::to::ToFFI;
@@ -135,7 +135,7 @@ impl FromFFI for types::MasternodeList {
                 BTreeMap::new(),
                 |mut acc, i| {
                     let llmq_map = *(*self.llmq_type_maps.add(i));
-                    let key = common::LLMQType::from(llmq_map.llmq_type);
+                    let key = chain::common::LLMQType::from(llmq_map.llmq_type);
                     let value: BTreeMap<UInt256, models::LLMQEntry> = (0..llmq_map.count)
                         .into_iter()
                         .fold(BTreeMap::new(), |mut acc, j| {
