+++ dash-spv-masternode-processor/src/ffi/to.rs
@@ -1,6 +1,6 @@
 use std::collections::BTreeMap;
 use std::ptr::null_mut;
-use crate::{common, models, tx, types};
+use crate::{chain, common, models, tx, types};
 use crate::crypto::UInt256;
 use crate::ffi::boxer::{boxed, boxed_vec};
 use crate::ffi::from::FromFFI;
@@ -332,7 +332,7 @@ impl ToFFI for models::QuorumsCLSigsObject {
 }
 
 pub fn encode_quorums_map(
-    quorums: &BTreeMap<common::LLMQType, BTreeMap<UInt256, models::LLMQEntry>>,
+    quorums: &BTreeMap<chain::common::LLMQType, BTreeMap<UInt256, models::LLMQEntry>>,
 ) -> *mut *mut types::LLMQMap {
     boxed_vec(
         quorums
