+++ dash-spv-masternode-processor/src/models/masternode_list.rs
@@ -1,11 +1,11 @@
 use std::cmp::min;
 use std::collections::BTreeMap;
-use crate::common::{LLMQType, MasternodeType};
+use crate::chain::common::LLMQType;
+use crate::common::MasternodeType;
+use crate::consensus::Encodable;
+use crate::crypto::{byte_util::{Reversable, Zeroable}, UInt256};
 use crate::models::{LLMQEntry, MasternodeEntry};
 use crate::tx::CoinbaseTransaction;
-use crate::consensus::Encodable;
-use crate::crypto::byte_util::{Reversable, Zeroable};
-use crate::crypto::UInt256;
 use crate::util::data_ops::merkle_root_from_hashes;
 
 #[derive(Clone, Ord, PartialOrd, Eq, PartialEq)]
