+++ dash-spv-masternode-processor/src/tests/json_from_core_snapshot.rs
@@ -3,7 +3,8 @@ use std::num::ParseIntError;
 use byte::BytesExt;
 use byte::ctx::Bytes;
 use serde::{Deserialize, Serialize};
-use crate::common::{LLMQSnapshotSkipMode, LLMQType, LLMQVersion, MasternodeType, SocketAddress};
+use crate::chain::common::LLMQType;
+use crate::common::{LLMQSnapshotSkipMode, LLMQVersion, MasternodeType, SocketAddress};
 use crate::consensus::encode::VarInt;
 use crate::crypto::{byte_util::{BytesDecodable, Reversable}, UInt160, UInt256, UInt384, UInt768, VarArray, VarBytes};
 use crate::hashes::hex::FromHex;
