+++ dash-spv-masternode-processor/src/tests/listdiff/llmq_verification.rs
@@ -5,8 +5,7 @@ use crate::bindings::masternode::process_mnlistdiff_from_message;
 use crate::{models, types};
 use crate::ffi::from::FromFFI;
 use crate::ffi::to::ToFFI;
-use crate::chain::common::chain_type::ChainType;
-use crate::common::LLMQType;
+use crate::chain::common::{ChainType, LLMQType};
 use crate::crypto::UInt256;
 use crate::lib_tests::tests::{add_insight_lookup_default, assert_diff_result, get_block_hash_by_height_default, get_llmq_snapshot_by_block_hash_default, get_masternode_list_by_block_hash_from_cache, get_merkle_root_by_hash_default, hash_destroy_default, masternode_list_destroy_default, masternode_list_save_in_cache, message_from_file, save_llmq_snapshot_default, should_process_diff_with_range_default, snapshot_destroy_default, FFIContext, get_block_height_by_hash_from_context};
 use crate::tests::block_store::init_testnet_store;
