+++ dash-spv-masternode-processor/src/tests/listdiff/quorum_issue.rs
@@ -1,4 +1,4 @@
-use crate::common::ChainType;
+use crate::chain::common::ChainType;
 use crate::lib_tests::tests::FFIContext;
 use crate::tests::block_store::init_mainnet_store;
 use crate::tests::listdiff::mainnet_reload::load_masternode_lists_for_files;
