+++ dash-spv-masternode-processor/src/tests/listdiff/test_basic_bls_scheme.rs
@@ -5,8 +5,7 @@ use hashes::hex::{FromHex, ToHex};
 use crate::{models, types};
 use crate::bindings::common::{processor_create_cache, register_processor};
 use crate::bindings::masternode::process_mnlistdiff_from_message;
-use crate::chain::common::chain_type::DevnetType;
-use crate::common::ChainType;
+use crate::chain::common::{ChainType, DevnetType};
 use crate::crypto::{UInt256, UInt384};
 use crate::crypto::byte_util::Zeroable;
 use crate::ffi::boxer::boxed;
