+++ dash-spv-masternode-processor/src/tests/listdiff/test_quorums_from_dash_core_snapshot.rs
@@ -1,6 +1,5 @@
 use crate::bindings::common::{processor_create_cache, register_processor};
-use crate::chain::common::chain_type::IHaveChainSettings;
-use crate::common::ChainType;
+use crate::chain::common::{ChainType, IHaveChainSettings};
 use crate::crypto::{byte_util::Reversable, UInt256};
 use crate::lib_tests::tests::{add_insight_lookup_default, FFIContext, get_block_hash_by_height_from_context, get_block_height_by_hash_from_context, get_llmq_snapshot_by_block_hash_default, get_masternode_list_by_block_hash_default, get_merkle_root_by_hash_default, hash_destroy_default, masternode_list_destroy_default, masternode_list_save_default, message_from_file, save_llmq_snapshot_default, should_process_diff_with_range_default, snapshot_destroy_default};
 use crate::tests::block_store::init_testnet_store;
